+++ packages/mars-build/src/compiler/script/script-h5.js
@@ -96,7 +96,7 @@ exports.compileMain = function (content, options) {
         color,
         selectedColor,
         list = []
-    } = tabBar;
+    } = tabBar || {};
     const tabBarStyle = {
         backgroundColor,
         borderStyle,